egl: Convert configs to use shifts instead of masks
Change dri2_add_config to accept an array of shifts instead of masks, and
compare with those set in the dri config. Convert all platform driver masks
to shifts.

In order to handle older drivers, where shift attributes aren't available,
we fall back to the mask attributes and compute the shifts with ffs.

Signed-off-by: Kevin Strasser <kevin.strasser@intel.com>
Reviewed-by: Adam Jackson <ajax@redhat.com>